WNB-181: Move routes folder to TS
