+++ package.json
@@ -44,6 +44,7 @@
     "@types/ramda": "0.26.36",
     "@types/react": "16.9.13",
     "@types/react-dom": "16.9.4",
+    "@types/react-loadable": "5.5.3",
     "@types/react-redux": "7.1.5",
     "@types/react-router-dom": "5.1.3",
     "@types/redux-actions": "2.6.1",
