+++ src/components/routes/ChangePassword/ChangePasswordForm/styles.js
@@ -1,5 +1,5 @@
 import styled, { keyframes } from 'styled-components';
-import { SubmitButton, Form, Error, PasswordInput } from 'components';
+import { SubmitButton, Form, Error, PasswordInput } from 'components/index';
 
 import { fontSmall } from 'ui-kit';
