+++ src/components/routes/ChangePassword/ChangePasswordSuccess/styles.js
