+++ src/components/routes/Profile/MyInfoPage/index.js
@@ -1,3 +1,3 @@
-export * from './ChangePassword';
+export * from '../../ChangePassword';
 export * from './MyInfoForm';
 export * from './MyInfoPage';
