+++ src/components/routes/index.ts
@@ -1,6 +1,7 @@
 export * from './Account';
 export * from './Categories';
 export * from './ForgotPassword';
+export * from './ChangePassword';
 export * from './Home';
 export * from './Profile';
 export * from './Signin';
