+++ src/routes/ChangePassword/ChangePassword.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+import { ChangePasswordForm } from 'components';
+
+const ChangePassword = () => {
+  return <ChangePasswordForm />;
+};
+
+export default ChangePassword;
