--- src/routes/Company/Company.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-import { OrganizationsList } from '../../components/OrganizationsList/index';
-
-export default class Company extends React.Component {
-  render() {
-    return <OrganizationsList />;
-  }
-}
