+++ src/routes/ForgotPassword/ForgotPassword.tsx
@@ -2,10 +2,10 @@ import React, { useContext } from 'react';
 import { useSelector } from 'react-redux';
 import { Redirect } from 'react-router-dom';
 
+import { ForgotPasswordForm, Gratitude } from 'components';
+
 import { AuthContext } from '../../components/App/contexts';
 import { getPageUrl } from '../RouteResolver';
-import { ForgotPasswordForm } from '../../components/routes/ForgotPassword/ForgotPasswordForm/ForgotPasswordForm';
-import { Gratitude } from '../../components/routes/ForgotPassword/Gratitude/Gratitude';
 
 import { ForgotPasswordContainer, Description } from './styles';
 import { getForgotPasswordLoadingProps } from '../../store/auth/forgotPassword';
