--- src/routes/Home/Home.js
@@ -1,15 +1,0 @@
-import React from 'react';
-
-import { MainSlider } from '../../components/routes/Home/MainSlider/index';
-import { AdvantagesList } from '../../components/routes/Home/AdvantagesList';
-
-const Home = () => {
-  return (
-    <>
-      <MainSlider />
-      <AdvantagesList />
-    </>
-  );
-};
-
-export default Home;
