+++ src/routes/Home/Home.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+import { MainSlider, AdvantagesList } from 'components';
+
+const Home = () => {
+  return (
+    <>
+      <MainSlider />
+      <AdvantagesList />
+    </>
+  );
+};
+
+export default Home;
