+++ src/routes/NotFound/NotFound.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
+
 import { getPageUrl } from '../RouteResolver';
 
 export default class NotFound extends React.Component {
