+++ src/routes/index.ts
@@ -1,11 +1,10 @@
 export * from './AppRoutes';
 export * from './RouteResolver';
-export * from './Account';
 export * from './Auth';
 export * from './Categories/Categories';
 export * from './Category/Category';
-export * from './Company/Company';
 export * from './ForgotPassword/ForgotPassword';
+export * from './ChangePassword/ChangePassword';
 export * from './Home/Home';
 export * from './NotFound/NotFound';
 export * from './Profile';
